Moved expression "parsing" to repl
